1. Added OpenAI OSS models, 2. Updated file structure adn its dependencies with new Manifest and init file system for pipx install detection and also for detection of other file types such as .json using manifest file, 3. Updated Readme file
