--- README.md
@@ -1,2 +1,0 @@
-# custom_cli
-Custom made cli using Groq platform
