--- groq_cli.egg-info/entry_points.txt
@@ -1,2 +1,0 @@
-[console_scripts]
-groq-cli = groq_cli:main
