--- groq_cli.egg-info/requires.txt
@@ -1,2 +1,0 @@
-groq
-requests
