+++ pyproject.toml
@@ -1,12 +1,16 @@
-
-
 [project]
 name = "groq-cli"
 version = "0.1.0"
 description = "Groq CLI"
 requires-python = ">=3.8"
-dependencies = ["groq","requests"]
+dependencies = ["groq", "requests"]
 
 [project.scripts]
-groq-cli = "groq_cli:main"
+groq-cli = "groq_cli_package.groq_cli:main"
+
+[tool.setuptools]
+include-package-data = true
+
+[tool.setuptools.packages.find]
+where = ["."]
